User-Agent addition
Good old weedmaps added PerimeterX to deter against web scraping. Adding the User-Agent header to the requests appears to get us back online!

Also rewrote a part of the getMenus function so it utilizes the same do_request function as the rest of the code.

Reach out if you see CAPTCHA challenges, can add functionality to open the browser to re-sume potentially.